% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{alarm_cache_size}
\alias{alarm_cache_size}
\alias{alarm_cache_clear}
\alias{alarm_cache_path}
\title{Work with the the \code{alarmdata} cache}
\usage{
alarm_cache_size()

alarm_cache_clear(force = FALSE)

alarm_cache_path()
}
\arguments{
\item{force}{FALSE by default. Asks the user to confirm if interactive. Does
not clear cache if force is FALSE and not interactive.}
}
\value{
For \code{alarm_cache_size()}, the size in bytes, invisibly

For \code{alarm_cache_clear()}, the path to the cache, invisibly.

For \code{alarm_cache_path()}, the path to the cache
}
\description{
Functions to inspect and clear the cache. If the cache is not enabled, uses a
temporary directory.
}
\examples{
alarm_cache_size()

alarm_cache_clear()

alarm_cache_path()

}
\concept{other}
