% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg.R
\name{alarm_epsg}
\alias{alarm_epsg}
\title{Suggested EPSG Codes}
\usage{
alarm_epsg(state)
}
\arguments{
\item{state}{A state name, abbreviation, FIPS code, or ANSI code}
}
\value{
numeric EPSG code
}
\description{
Provides suggested EPSG codes for each of the 50 states.
}
\examples{
alarm_epsg('NY')
}
