% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmdata-package.R
\docType{package}
\name{alarmdata-package}
\alias{alarmdata}
\alias{alarmdata-package}
\title{alarmdata: Download, Merge, and Process Redistricting Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to download and process data produced by the ALARM Project, including 2020 redistricting files Kenny and McCartan (2021) \url{https://alarm-redist.org/posts/2021-08-10-census-2020/} and the 50-State Redistricting Simulations of McCartan, Kenny, Simko, Garcia, Wang, Wu, Kuriwaki, and Imai (2022) \doi{10.7910/DVN/SLCD3E}. The package extends the data introduced in McCartan, Kenny, Simko, Garcia, Wang, Wu, Kuriwaki, and Imai (2022) \doi{10.1038/s41597-022-01808-2} to also include states with only a single district. The package also includes the Japanese 2022 redistricting files from the 47-Prefecture Redistricting Simulations of Miyazaki, Yamada, Yatsuhashi, and Imai (2022) \doi{10.7910/DVN/Z9UKSH}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/alarmdata/}
  \item \url{https://alarm-redist.org/alarmdata/}
  \item Report bugs at \url{https://github.com/alarm-redist/alarmdata/issues/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

Authors:
\itemize{
  \item Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})
  \item Tyler Simko \email{tsimko@g.harvard.edu} (\href{https://orcid.org/0000-0001-9382-7346}{ORCID})
  \item Michael Zhao \email{michaelzhao@college.harvard.edu} (\href{https://orcid.org/0000-0001-8435-2078}{ORCID})
  \item Sho Miyazaki \email{shomiya@stanford.edu} (\href{https://orcid.org/0009-0009-7838-9005}{ORCID})
  \item Brian Zhou \email{hello@zhoubrian.com} (\href{https://orcid.org/0000-0002-0957-5965}{ORCID})
  \item Kosuke Imai \email{imai@harvard.edu} (\href{https://orcid.org/0000-0002-2748-1022}{ORCID})
}

}
\keyword{internal}
