% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiftystates.R
\name{alarm_50state}
\alias{alarm_50state}
\alias{alarm_50state_map}
\alias{alarm_50state_plans}
\title{Download maps and plans from the 50-State Simulation Project}
\usage{
alarm_50state_map(state)

alarm_50state_plans(state)
}
\arguments{
\item{state}{A state name, abbreviation, FIPS code, or ANSI code}
}
\description{
These functions will download \link{redist_map} and \link{redist_plans} objects for the
50-State Simulation Project from the ALARM Project's Dataverse.
}
\details{
Every decade following the Census, states and municipalities must redraw
districts for Congress, state houses, city councils, and more. The goal of
the 50-State Simulation Project is to enable researchers, practitioners, and
the general public to use cutting-edge redistricting simulation analysis to
evaluate enacted congressional districts.

Evaluating a redistricting plan requires analysts to take into account each
state’s redistricting rules and particular political geography. Comparing the
partisan bias of a plan for Texas with the bias of a plan for New York, for
example, is likely misleading. Comparing a state’s current plan to a past
plan is also problematic because of demographic and political changes over
time. Redistricting simulations generate an ensemble of alternative
redistricting plans within a given state which are tailored to its
redistricting rules. Unlike traditional evaluation methods, therefore,
simulations are able to directly account for the state’s political geography
and redistricting criteria.
}
