% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmdata-package.R
\docType{package}
\name{alarmdata-package}
\alias{alarmdata}
\alias{alarmdata-package}
\title{alarmdata: Download, Merge, and Process Redistricting Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to download and process data produced by the ALARM Project, including 2020 redistricting files \url{https://alarm-redist.org/posts/2021-08-10-census-2020/} and 50-State Redistricting Simulations \doi{10.7910/DVN/SLCD3E}. The package extends functionality introduced in \href{https://arxiv.org/abs/2206.10763}{arXiv:2206.10763} to cover states with only a single district.
}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

Authors:
\itemize{
  \item Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})
  \item Tyler Simko \email{tsimko@g.harvard.edu}
  \item Michael Zhao \email{michaelzhao@college.harvard.edu} (\href{https://orcid.org/0000-0001-8435-2078}{ORCID})
  \item Kosuke Imai \email{imai@harvard.edu}
}

}
\keyword{internal}
