% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmdata-package.R
\docType{package}
\name{alarmdata-package}
\alias{alarmdata}
\alias{alarmdata-package}
\title{alarmdata: Download, Merge, and Process Redistricting Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Utility functions to download and process data produced by the ALARM Project, including 2020 redistricting files <https://alarm-redist.github.io/posts/2021-08-10-census-2020/> and 50-State Redistricting Simulations <doi:10.7910/DVN/SLCD3E>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alarm-redist.github.io/alarmdata/}
  \item Report bugs at \url{https://github.com/alarm-redist/alarmdata/issues/}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

Authors:
\itemize{
  \item Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})
  \item Tyler Simko \email{tsimko@g.harvard.edu}
  \item Michael Zhao \email{michaelzhao@college.harvard.edu}
  \item Kosuke Imai \email{imai@harvard.edu}
}

}
\keyword{internal}
